requirements.txt
langchain==0.0.222
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.24.0
openai==0.27.8
faiss-cpu==1.7.4
tiktoken
huggingface-hub
InstructorEmbedding
sentence-transformers
redis
